+++ code/system/sys.data.crdt/src/test/-TEST.v.mts
@@ -1,4 +1,5 @@
-import { Test, describe, it, type t } from '.';
+import { Test, describe, expect, it, type t } from '.';
+import { getTestFilesystem } from '../test.ui';
 import { TESTS } from '../test.ui/-TestRunner.TESTS.mjs';
 
 /**
@@ -21,5 +22,10 @@ describe('visual specs', () => {
     const { Specs } = await import('../test.ui/entry.Specs.mjs');
     const res = await Dev.headless(Specs);
     expect(res.ok).to.eql(true);
+  });
+
+  it('test filesystem', async (e) => {
+    const fs = await getTestFilesystem();
+    expect(fs.events.id).to.match(/^fs\.memory\./); // NB: because running on node.
   });
 });
